workers/pyproject.toml
[tool.poetry]
name = "ohsome_quality_analyst"
version = "0.9.0"
description = "Intrinsic and extrinsic data quality metrics for OpenStreetMap data."
authors = ["ohsome team <ohsome@heigit.org>"]
homepage = "https://oqt.ohsome.org"
repository = "https://github.com/GIScience/ohsome-quality-analyst"
keywords = [
  "osm",
  "gis",
  "geography",
  "metric",
  "ohsome",
  "quality",
  ]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"  # `rasterstats` restricts Python version to below 3.10
geojson = "^2.5.0"
fastapi = "^0.68.0"
requests = "^2.25.0"
matplotlib = "^3.2.2"
dacite = "^1.6.0"
PyYAML = "^5.4.1"
toml = "^0.10.2"
httpx = "^0.17.0"
asyncpg = "^0.25.0"
vcrpy = "^4.1.1"
python-dateutil = "^2.8.2"
Jinja2 = "^3.0.3"
scipy = "^1.7.3"  # `scipy` requires Python >=3.7,<3.11
rpy2 = "^3.4.5"
scikit-learn = "^1.0.2"  # Only works on Python below 3.11 (Depends on `scipy`)
rasterstats = "^0.16.0"  # Only works on Python below 3.10 (Depends on `rasterio`)
pyproj = "^3.3.0"
click = "^8.1.2"
uvicorn = "^0.17.6"
schema = "^0.7.5"

[tool.poetry.dev-dependencies]
mypy = "0.930"
pytest = "^6.2.2"
pytest-cov = "^2.12.0"
pep8-naming = "^0.12.1"
pre-commit = "^2.15.0"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.poetry.scripts]
oqt = "ohsome_quality_analyst.cli.cli:cli"

[tool.poetry.urls]
"Issues" = "https://github.com/GIScience/ohsome-quality-analyst/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
